+++ tests/common/mod.rs
@@ -37,7 +37,7 @@ macro_rules! test_node {
             name: Some($name.into()),
             ..Default::default()
         };
-        common::TestNode(pea2pea::Node::new(Some(config)).await.unwrap())
+        common::TestNode(pea2pea::Node::new(config).await.unwrap())
     }};
 }
 
@@ -45,7 +45,7 @@ pub async fn start_test_nodes(count: usize) -> Vec<TestNode> {
     let mut nodes = Vec::with_capacity(count);
 
     for _ in 0..count {
-        nodes.push(TestNode(Node::new(None).await.unwrap()));
+        nodes.push(TestNode(Node::new(Default::default()).await.unwrap()));
     }
 
     nodes