+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+- `Node::new` now takes `Config` instead of `Option<Config>`
 - renamed `Writing::{send_direct_message, send_broadcast}` to `::{unicast, broadcast}`
 
 # 0.39.0